zrpcd: change BGP_ETHTAG_MAX_ET from i32 to i64 thrift value
thrift value for BGP_ETHTAG_MAX_ET is set to i64 value.

+++ zrpcd/vpnservice.dualstack.thrift
@@ -44,7 +44,7 @@
  const i32 BGP_ERR_PARAM = 100
  const i32 BGP_ERR_NOT_SUPPORTED = 200
  
- const i32 BGP_ETHTAG_MAX_ET = 0xffffffff
+ const i64 BGP_ETHTAG_MAX_ET = 0xffffffff
 
  // supported afi-safi combinations 
  enum af_afi {
